backend/webapp/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
bidict==0.22.1
certifi==2024.2.2
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.2.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
python-engineio==4.9.0
python-socketio==5.11.1
replicate==0.23.1
simple-websocket==1.0.0
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
torch==2.2.0
triton==2.2.0
typing_extensions==4.9.0
uvicorn==0.27.0.post1
wsproto==1.2.0
